requirements.txt
flask
python-telegram-bot
requests
beautifulsoup4
python-dotenv
selenium
webdriver-manager
apscheduler

